// src/generated/types/EventDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type EventDto = {
  /**
   * @type string
   */
  name: string
  /**
   * @type number
   */
  duration: number
  /**
   * @type boolean
   */
  isExam: boolean
  /**
   * @type boolean
   */
  isInLab: boolean
  /**
   * @type string | undefined
   */
  examFiles?: string | undefined
  /**
   * @type number
   */
  degree: number
  /**
   * @default false
   * @type boolean
   */
  autoStart: boolean
  /**
   * @default 30
   * @type number
   */
  examModeStartMinutes: number
  /**
   * @type string | undefined
   */
  description?: string | undefined
  /**
   * @type string
   */
  courseId: string
  /**
   * @type string
   */
  id: string
}